requirements.txt
# Core ML Libraries
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0
open-clip-torch>=2.20.0

# Data Processing
pandas>=2.0.0
pyarrow>=12.0.0
numpy>=1.24.0
pillow>=10.0.0

# API
openai>=1.0.0
aiohttp>=3.8.0

# Visualization
matplotlib>=3.7.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
pydantic>=2.0.0

# Optional: For better performance
accelerate>=0.20.0

